+++ crd-generator/api-v2/src/main/java/io/fabric8/crdv2/generator/v1/JsonSchema.java
@@ -24,17 +24,22 @@ import io.fabric8.crdv2.generator.v1.JsonSchema.V1JSONSchemaProps;
 import io.fabric8.crdv2.generator.v1.JsonSchema.V1ValidationRule;
 import io.fabric8.generator.annotation.JSONSchema;
 import io.fabric8.kubernetes.api.model.apiextensions.v1.ExternalDocumentation;
-import io.fabric8.kubernetes.api.model.apiextensions.v1.ExternalDocumentationBuilder;
 import io.fabric8.kubernetes.api.model.apiextensions.v1.JSONSchemaProps;
+import io.fabric8.kubernetes.api.model.apiextensions.v1.JSONSchemaPropsBuilder;
 import io.fabric8.kubernetes.api.model.apiextensions.v1.JSONSchemaPropsOrArray;
 import io.fabric8.kubernetes.api.model.apiextensions.v1.JSONSchemaPropsOrBool;
 import io.fabric8.kubernetes.api.model.apiextensions.v1.JSONSchemaPropsOrStringArray;
 import io.fabric8.kubernetes.api.model.apiextensions.v1.ValidationRule;
 
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -117,57 +122,99 @@ public class JsonSchema extends AbstractJsonSchema<V1JSONSchemaProps, V1Validati
 
   @Override
   protected V1JSONSchemaProps fromAnnotation(Class<?> rawClass, boolean isTargetType, JSONSchema schema) {
-      V1JSONSchemaProps result = super.fromAnnotation(rawClass, isTargetType, schema);
-      // maybe override the type if it was determined by reading the optional `implementation`
-      setIfDefined(mapDefined(schema.type()), result::setType);
-      setIfDefined(mapDefined(schema.$ref()), result::set$ref);
-      setIfDefined(mapDefined(schema.$schema()), result::set$schema);
-      setIfDefined(mapSchemaOrBool(schema.additionalItems()), result::setAdditionalItems);
-      setIfDefined(mapSchemaOrBool(schema.additionalProperties()), result::setAdditionalProperties);
-      setIfDefined(mapSchemaList(schema.allOf()), result::setAllOf);
-      setIfDefined(mapSchemaList(schema.anyOf()), result::setAnyOf);
-      setIfDefined(mapSchemaMap(schema.definitions()), result::setDefinitions);
-      setIfDefined(mapDependencies(schema.dependencies()), result::setDependencies);
-      setIfDefined(mapEnumeration(schema.enumeration(), rawClass), result::setEnum);
-      setIfDefined(mapDefined(schema.example(), rawClass), result::setExample);
-      setIfDefined(mapExternalDocs(schema.externalDocs()), result::setExternalDocs);
-      setIfDefined(mapDefined(schema.id()), result::setId);
-      setIfDefined(mapSchemaOrArray(schema.items()), result::setItems);
-      setIfDefined(mapDefined(schema.multipleOf()), result::setMultipleOf);
-      setIfDefined(mapSchema(schema.not()), result::setNot);
-      setIfDefined(mapSchemaList(schema.oneOf()), result::setOneOf);
-      setIfDefined(mapSchemaMap(schema.patternProperties()), result::setPatternProperties);
-      setIfDefined(mapSchemaMap(schema.properties()), result::setProperties);
-      setIfDefined(mapDefined(schema.title()), result::setTitle);
-      setIfDefined(mapBoolean(schema.uniqueItems()), result::setUniqueItems);
-      setIfDefined(mapBoolean(schema.xKubernetesEmbeddedResource()), result::setXKubernetesEmbeddedResource);
-      setIfDefined(mapBoolean(schema.xKubernetesIntOrString()), result::setXKubernetesIntOrString);
-      setIfDefined(mapDefined(schema.xKubernetesListMapKeys()), result::setXKubernetesListMapKeys);
-      setIfDefined(mapDefined(schema.xKubernetesListType()), result::setXKubernetesListType);
-      setIfDefined(mapDefined(schema.xKubernetesMapType()), result::setXKubernetesMapType);
-      setIfDefined(mapValidationRules(schema.xKubernetesValidations()), result::setXKubernetesValidations);
-      return result;
+    V1JSONSchemaProps result = super.fromAnnotation(rawClass, isTargetType, schema);
+    // maybe override the type if it was determined by reading the optional `implementation`
+    setIfDefined(schema.type(), result::setType);
+    setIfDefined(schema.$ref(), result::set$ref);
+    setIfDefined(schema.$schema(), result::set$schema);
+    setIfDefined(schema.additionalItems(), this::mapSchemaOrBool, result::setAdditionalItems);
+    setIfDefined(schema.additionalProperties(), this::mapSchemaOrBool, result::setAdditionalProperties);
+    setIfDefined(schema.definitions(), LinkedHashMap::new, this::mapSchemaMap, result::setDefinitions);
+    setIfDefined(schema.dependencies(), LinkedHashMap::new, this::mapDependencies, result::setDependencies);
+    setIfDefined(schema.enumeration(), ArrayList::new, v -> parseJson(v, rawClass), result::setEnum);
+    setIfDefined(schema.example(), v -> parseJson(v, rawClass), result::setExample);
+    setIfDefined(schema.externalDocs(), this::mapExternalDocs, result::setExternalDocs);
+    setIfDefined(schema.id(), result::setId);
+    setIfDefined(schema.multipleOf(), result::setMultipleOf);
+    setIfDefined(schema.patternProperties(), LinkedHashMap::new, this::mapSchemaMap, result::setPatternProperties);
+    setIfDefined(schema.title(), result::setTitle);
+    setIfDefined(schema.uniqueItems(), this::mapBoolean, result::setUniqueItems);
+    setIfDefined(schema.xKubernetesEmbeddedResource(), this::mapBoolean, result::setXKubernetesEmbeddedResource);
+    setIfDefined(schema.xKubernetesIntOrString(), this::mapBoolean, result::setXKubernetesIntOrString);
+    setIfDefined(schema.xKubernetesListMapKeys(), ArrayList::new, Arrays::asList, result::setXKubernetesListMapKeys);
+    setIfDefined(schema.xKubernetesListType(), result::setXKubernetesListType);
+    setIfDefined(schema.xKubernetesMapType(), result::setXKubernetesMapType);
+    setIfDefined(schema.xKubernetesValidations(), ArrayList::new, this::mapValidationRules, result::setXKubernetesValidations);
+
+    if (schema.structural()) {
+      List<JSONSchemaProps> allOf = new ArrayList<>(schema.allOf().length);
+      List<JSONSchemaProps> anyOf = new ArrayList<>(schema.anyOf().length);
+      AtomicReference<JSONSchemaPropsOrArray> items = new AtomicReference<>();
+      AtomicReference<JSONSchemaProps> not = new AtomicReference<>();
+      List<JSONSchemaProps> oneOf = new ArrayList<>(schema.oneOf().length);
+      Map<String, JSONSchemaProps> properties = new LinkedHashMap<>();
+
+      Consumer<JSONSchemaProps> addProperties = s -> s.getProperties().forEach(properties::putIfAbsent);
+      /**
+       * Only sets the items if they have not yet been set.
+       */
+      Consumer<JSONSchemaProps> maybeSetItems = s -> items.compareAndSet(null, s.getItems());
+
+      setIfDefined(schema.items(), this::mapSchemaOrArray, items::set);
+
+      setIfDefined(schema.allOf(), ArrayList::new, this::mapSchemaList, (List<JSONSchemaProps> schemas) -> {
+        schemas.stream().map(this::mapJunctorSchema).forEach(allOf::add);
+        schemas.forEach(addProperties);
+        schemas.forEach(maybeSetItems);
+      });
+
+      setIfDefined(schema.anyOf(), ArrayList::new, this::mapSchemaList, (List<JSONSchemaProps> schemas) -> {
+        schemas.stream().map(this::mapJunctorSchema).forEach(anyOf::add);
+        schemas.forEach(addProperties);
+        schemas.forEach(maybeSetItems);
+      });
+
+      setIfDefined(schema.not(), this::mapSchema, notSchema -> {
+        not.set(mapJunctorSchema(notSchema));
+        addProperties.accept(notSchema);
+        maybeSetItems.accept(notSchema);
+      });
+
+      setIfDefined(schema.oneOf(), ArrayList::new, this::mapSchemaList, (List<JSONSchemaProps> schemas) -> {
+        schemas.stream().map(this::mapJunctorSchema).forEach(oneOf::add);
+        schemas.forEach(addProperties);
+        schemas.forEach(maybeSetItems);
+      });
+
+      setIfDefined(schema.properties(), LinkedHashMap::new, this::mapSchemaMap, properties::putAll);
+
+      result.setAllOf(allOf);
+      result.setAnyOf(anyOf);
+      result.setItems(items.get());
+      result.setNot(not.get());
+      result.setOneOf(oneOf);
+      result.setProperties(properties);
+    } else {
+      setIfDefined(schema.allOf(), ArrayList::new, this::mapSchemaList, result::setAllOf);
+      setIfDefined(schema.anyOf(), ArrayList::new, this::mapSchemaList, result::setAnyOf);
+      setIfDefined(schema.items(), this::mapSchemaOrArray, result::setItems);
+      setIfDefined(schema.not(), this::mapSchema, result::setNot);
+      setIfDefined(schema.oneOf(), ArrayList::new, this::mapSchemaList, result::setOneOf);
+      setIfDefined(schema.properties(), LinkedHashMap::new, this::mapSchemaMap, result::setProperties);
+    }
+
+    return result;
   }
 
   @Override
   protected V1JSONSchemaProps mapImplementation(Class<?> value, boolean isTargetType) {
-    if (value == JSONSchema.Undefined.class) {
+    if (JSONSchema.Undefined.isUndefined(value)) {
       return null; // NOSONAR
     }
     return new JsonSchema(resolvingContext.forkContext(isTargetType), value).getSchema();
   }
 
-  private JSONSchemaProps mapSchema(Class<?> value) {
-    if (value == JSONSchema.Undefined.class) {
-      return null; // NOSONAR
-    }
-    return new JsonSchema(resolvingContext.forkContext(false), value).getSchema();
-  }
-
   private JSONSchemaPropsOrBool mapSchemaOrBool(Class<?> value) {
-    if (value == JSONSchema.Undefined.class) {
-      return null; // NOSONAR
-    }
     JSONSchemaPropsOrBool result = new JSONSchemaPropsOrBool();
 
     if (JSONSchema.Boolean.class.isAssignableFrom(value)) {
@@ -181,66 +228,89 @@ public class JsonSchema extends AbstractJsonSchema<V1JSONSchemaProps, V1Validati
     return result;
   }
 
-  private List<JsonNode> mapEnumeration(String[] examples, Class<?> targetType) {
-    if (examples.length != 0) {
-      return Arrays.stream(examples).map(ex -> mapDefined(ex, targetType)).collect(Collectors.toList());
-    }
-    return null; // NOSONAR
+  private JSONSchemaPropsOrArray mapSchemaOrArray(Class<?>[] values) {
+    return mapSchemaOrArray(mapSchemaList(values));
   }
 
-  private JSONSchemaPropsOrArray mapSchemaOrArray(Class<?>[] values) {
-    return Optional.ofNullable(mapSchemaList(values))
-        .map(schemas -> {
-          JSONSchemaPropsOrArray result = new JSONSchemaPropsOrArray();
-          if (schemas.size() == 1) {
-            result.setSchema(schemas.get(0));
-          } else {
-            result.setJSONSchemas(schemas);
-          }
-          return result;
-        })
-        .orElse(null);
+  private JSONSchemaPropsOrArray mapSchemaOrArray(List<JSONSchemaProps> schemas) {
+    JSONSchemaPropsOrArray result = new JSONSchemaPropsOrArray();
+
+    if (schemas.size() == 1) {
+      result.setSchema(schemas.get(0));
+    } else {
+      result.setJSONSchemas(schemas);
+    }
+
+    return result;
+  }
+
+  private JSONSchemaProps mapSchema(Class<?> value) {
+    return new JsonSchema(resolvingContext.forkContext(false), value).getSchema();
   }
 
   private List<JSONSchemaProps> mapSchemaList(Class<?>[] values) {
-    if (values.length == 0) {
-      return null; // NOSONAR
+    return Arrays.stream(values).map(this::mapSchema).collect(Collectors.toList());
+  }
+
+  /**
+   * Suppress attributes disallowed within a logical junctor (allOf, anyOf, oneOf, not). This functionality
+   * is in support of rule #3 from "Specifying a structural schema".
+   *
+   * @see https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definitions/#specifying-a-structural-schema
+   */
+  private JSONSchemaProps mapJunctorSchema(JSONSchemaProps schema) {
+    JSONSchemaPropsBuilder builder = schema.edit()
+        .withDescription(null)
+        .withDefault(null)
+        .withAdditionalProperties(null)
+        .withNullable(null)
+        .withItems(Optional.ofNullable(schema.getItems())
+            .map(items -> Stream
+                .concat(
+                    Optional.ofNullable(items.getJSONSchemas())
+                        .map(List::stream)
+                        .orElseGet(Stream::empty),
+                    Stream.of(items.getSchema()))
+                .filter(Objects::nonNull)
+                .map(this::mapJunctorSchema)
+                .collect(Collectors.toList()))
+            .map(this::mapSchemaOrArray)
+            .orElse(null))
+        .withProperties(schema.getProperties().entrySet()
+            .stream()
+            // Recursively remove property attributes
+            .map(e -> Map.entry(e.getKey(), mapJunctorSchema(e.getValue())))
+            .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue)));
+
+    if (!Boolean.TRUE.equals(schema.getXKubernetesIntOrString())) {
+      builder.withType(null);
     }
 
-    return Arrays.stream(values)
-        .map(this::mapSchema)
-        .collect(Collectors.toList());
+    return builder.build();
   }
 
   private Map<String, JSONSchemaProps> mapSchemaMap(JSONSchema.Map[] entries) {
-    if (entries.length == 0) {
-      return null; // NOSONAR
-    }
-
     return Arrays.stream(entries)
         .map(e -> Map.entry(e.name(), mapSchema(e.value())))
         .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
   }
 
   private Map<String, JSONSchemaPropsOrStringArray> mapDependencies(JSONSchema.DependencyMap[] entries) {
-    if (entries.length == 0) {
-      return null; // NOSONAR
-    }
-
     return Arrays.stream(entries)
-        .map(e -> Map.entry(e.name(), new JSONSchemaPropsOrStringArray(mapDefined(e.value().properties()), mapSchema(e.value().schema()))))
+        .map(e -> {
+          JSONSchemaPropsOrStringArray result = new JSONSchemaPropsOrStringArray();
+          setIfDefined(e.value().properties(), ArrayList::new, Arrays::asList, result::setProperty);
+          setIfDefined(e.value().schema(), this::mapSchema, result::setSchema);
+          return Map.entry(e.name(), result);
+        })
         .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
   }
 
   private ExternalDocumentation mapExternalDocs(JSONSchema.ExternalDocumentation externalDocs) {
-    if (Stream.of(externalDocs.description(), externalDocs.url()).allMatch(JSONSchema.Undefined.STRING::equals)) {
-      return null;
-    }
-
-    return new ExternalDocumentationBuilder()
-        .withDescription(mapDefined(externalDocs.description()))
-        .withUrl(mapDefined(externalDocs.url()))
-        .build();
+    ExternalDocumentation result = new ExternalDocumentation();
+    setIfDefined(externalDocs.description(), result::setDescription);
+    setIfDefined(externalDocs.url(), result::setUrl);
+    return result;
   }
 
   private List<ValidationRule> mapValidationRules(io.fabric8.generator.annotation.ValidationRule[] values) {